@@ -13,15 +13,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.10",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
+    "@angular/animations": "^6.1.10",
+    "@angular/common": "^6.1.10",
+    "@angular/compiler": "^6.1.10",
+    "@angular/core": "^6.1.10",
+    "@angular/forms": "^6.1.10",
+    "@angular/http": "^6.1.10",
+    "@angular/platform-browser": "^6.1.10",
+    "@angular/platform-browser-dynamic": "^6.1.10",
+    "@angular/router": "^6.1.10",
     "@ng-bootstrap/ng-bootstrap": "~1.1.2",
     "@progress/kendo-angular-buttons": "^4.0.0",
     "@progress/kendo-angular-dateinputs": "^3.1.0",
@@ -43,16 +43,17 @@
     "mdbootstrap": "^4.5.0",
     "moment": "^2.24.0",
     "oidc-client": "^1.3.0",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.0.0",
     "zone.js": "^0.8.19"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular/cli": "^6.0.1",
+    "@angular/compiler-cli": "^6.1.10",
+    "@angular/language-service": "^6.1.10",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
+    "babel-polyfill": "^6.23.0",
     "codelyzer": "^4.0.1",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
@@ -63,9 +64,9 @@
     "karma-jasmine-html-reporter": "^0.2.2",
     "node-sass": "^4.12.0",
     "protractor": "~5.1.2",
+    "rxjs-tslint": "^0.1.8",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3",
-    "babel-polyfill": "^6.23.0"
+    "typescript": "~2.5.3"
   }
 }
